requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.2
tqdm>=4.66.1
colorama>=0.4.6
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
textblob>=0.17.1
scipy>=1.10.0
ipython>=8.12.0
Flask>=2.0.0
black>=23.12.1
flake8>=4.0.1
altair>=5.2.0
vegafusion>=1.4.0
vegafusion-python-embed>=1.4.0
vl-convert-python>=1.6.0